refactor: move i18n methods, they must only be used in beans
Moved many static methods from UtilBean to ApplicationBean so that they
are only accessible in beans. Some methods could not be deleted but only
marked as deprecated because they are still used by many classes. This
has to be changed step by step.
The same applies to getUserBean. It must never be called from model
classes.